Update scruit to avoir ssl renewal each time

+++ start_containers.sh
@@ -2,7 +2,7 @@
 set -e
 
 # Create required directories
-mkdir certbot nginx superset_home
+mkdir -p certbot nginx superset_home
 
 # Step 1: Copy the HTTP configuration to the nginx folder
 echo "-- > Copying the HTTP configuration to nginx/nginx.conf"
@@ -12,20 +12,26 @@ cp nginx/nginx.http.conf nginx/nginx.conf
 echo "Starting nginx only in HTTP mode"
 docker-compose up -d nginx
 
-# Step 3: Run Certbot to obtain the certificates
-echo "-- > Running Certbot to obtain the certificates"
-docker run --rm -it \
-  -v "$(pwd)/certbot/www:/var/www/certbot" \
-  -v "$(pwd)/certbot/conf:/etc/letsencrypt" \
-  certbot/certbot certonly --webroot --webroot-path=/var/www/certbot \
-  --email tu-email@dominio.com --agree-tos --no-eff-email \
-  -d happyrobot-challenge.duckdns.org -d www.happyrobot-challenge.duckdns.org
+# Step 3: Ask if you want to create or renew certificates
+read -p "¿Want to create or renew SSL certificates? (s/n): " cert_choice
+if [[ "$cert_choice" =~ ^[Ss]$ ]]; then
+  echo "-- > Running Certbot to obtain the certificates"
+  docker run --rm -it \
+    -v "$(pwd)/certbot/www:/var/www/certbot" \
+    -v "$(pwd)/certbot/conf:/etc/letsencrypt" \
+    certbot/certbot certonly --webroot --webroot-path=/var/www/certbot \
+    --email tu-email@dominio.com --agree-tos --no-eff-email \
+    -d happyrobot-challenge.duckdns.org -d www.happyrobot-challenge.duckdns.org
 
-if [ $? -ne 0 ]; then
-  echo "-- > ERROR: Certbot failed. Exiting."
-  exit 1
+  if [ $? -ne 0 ]; then
+    echo "-- > ERROR: Certbot failed. Exiting."
+    exit 1
+  fi
+else
+  echo "-- > Skiping SSL certificate creation"
 fi
-# Step 4: "Start Superset dashboard and create admin user"
+
+# Step 4: Start Superset dashboard and create admin user
 echo "-- > Start Superset dashboard and create admin user"
 docker-compose up -d superset
 docker exec -it superset superset db upgrade
